+++ NEWS.md
@@ -1,5 +1,7 @@
-# codeInspectr (development version)
+# codeInspectr 0.0.0.9000 (development version)
 
+* Added `backup_github_info()` to download Issues and Pull Requests to local jsons.
+* added better input validation
 * addition of basic functions. Key ones: `plot_function_dependencies()` and `find_reverse_dependencies()`
 * Suite of functions to summarize information about github repositories
 * added vignette "getting_started"